requirements.txt
# Core RAG & Vector Store
langchain==0.3.25
langchain-community==0.3.23
faiss-cpu==1.9.0
sentence-transformers==3.4.1

# LLM APIs
groq==0.23.1
google-generativeai==0.8.5

# Document Parsing
pypdf==5.4.0
python-docx==1.1.2
pyyaml==6.0.2

# Frontend
streamlit==1.44.1

# Evaluation
pytest==8.3.5
pytest-cov==6.1.0

# Utilities
python-dotenv==1.1.0